Extract struct parser_priv into separate header file
